Fixed link to dcos.io
I know that this repository is deprecated, but at least the link to the successor repository should work fine. However, I have found, that the current link "dcos.io" link leads to "https://github.com/mesosphere/open-docs/blob/master/dcos.io" returning a "404 Not found" message.

Did you mean to point to "https://dcos.io/docs" or "https://dcos.io/" instead? If, so, I propose to change the link accordingly (I guess, the docs link is better, but it is your choice). 

P.S.: It seems, that there is no real successor of this project? https://dcos.io/docs is not a community based documentation, right?